requirements.txt
requests == 2.30.0
google-cloud-bigquery == 3.10.0
google-cloud-storage == 2.10.0